fix sergey's comments
- Add new formats to wrappers.
- Remove redundant resolution data from motion frame allocation.
- Restore image-avx.
