+++ src/proc/synthetic-stream.h
@@ -29,8 +29,6 @@ namespace librealsense
 
         frame_interface* allocate_motion_frame(std::shared_ptr<stream_profile_interface> stream,
             frame_interface* original,
-            int width = 0,
-            int height = 0,
             rs2_extension frame_type = RS2_EXTENSION_MOTION_FRAME) override;
 
         frame_interface* allocate_composite_frame(std::vector<frame_holder> frames) override;
