+++ wrappers/nodejs/index.js
@@ -3914,6 +3914,33 @@ const format = {
    * <br>Equivalent to its uppercase counterpart.
    */
   format_disparity32: 'disparity32',
+   /**
+   * String literal of <code>'mjpeg'</code>.
+   * <br>Bitstream encoding for video in which an image of each frame is encoded as JPEG-DIB.
+   */
+  format_mjpeg: 'mjpeg',
+   /**
+   * String literal of <code>'y8i'</code>. <br>8-bit per pixel interleaved. 8-bit left, 8-bit right.
+   */
+  format_y8i: 'y8i',
+   /**
+   * String literal of <code>'y12i'</code>.
+   * <br>12-bit per pixel interleaved. 12-bit left, 12-bit right. Each pixel is stored in a 24-bit word in little-endian order.
+   */
+  format_y12i: 'y12i',
+   /**
+   * String literal of <code>'inzi'</code>. <br>multi-planar Depth 16bit + IR 10bit.
+   */
+  format_inzi: 'inzi',
+   /**
+   * String literal of <code>'invi'</code>. <br>8-bit IR stream
+   */
+  format_invi: 'invi',
+   /**
+   * String literal of <code>'w10'</code>.
+   * <br>Grey-scale image as a bit-packed array. 4 pixel data stream taking 5 bytes
+   */
+  format_w10: 'w10',
   /**
    * When passed to enable stream, librealsense will try to provide best suited
    * format. <br>Equivalent to its lowercase counterpart.
@@ -4029,6 +4056,36 @@ const format = {
    * @type {Integer}
    */
   FORMAT_DISPARITY32: RS2.RS2_FORMAT_DISPARITY32,
+   /**
+   * Bitstream encoding for video in which an image of each frame is encoded as JPEG-DIB.
+   * @type {Integer}
+   */
+  FORMAT_MJPEG: RS2.RS2_FORMAT_MJPEG,
+   /**
+   * 8-bit per pixel interleaved. 8-bit left, 8-bit right.
+   * @type {Integer}
+   */
+  FORMAT_Y8I: RS2.RS2_FORMAT_Y8I,
+   /**
+   * 12-bit per pixel interleaved. 12-bit left, 12-bit right. Each pixel is stored in a 24-bit word in little-endian order.
+   * @type {Integer}
+   */
+  FORMAT_Y12I: RS2.RS2_FORMAT_Y12I,
+   /**
+   * multi-planar Depth 16bit + IR 10bit.
+   * @type {Integer}
+   */
+  FORMAT_INZI: RS2.RS2_FORMAT_INZI,
+   /**
+   * 8-bit IR stream.
+   * @type {Integer}
+   */
+  FORMAT_INVI: RS2.RS2_FORMAT_INVI,
+   /**
+   * Grey-scale image as a bit-packed array. 4 pixel data stream taking 5 bytes.
+   * @type {Integer}
+   */
+  FORMAT_W10: RS2.RS2_FORMAT_W10,
   /**
    * Number of enumeration values. Not a valid input: intended to be used in for-loops.
    * <br>Equivalent to its lowercase counterpart.
@@ -4085,7 +4142,19 @@ const format = {
       case this.FORMAT_6DOF:
         return this.format_6dof;
       case this.FORMAT_DISPARITY32:
-        return this.format_disparity32;
+		return this.format_disparity32;
+	  case this.RS2_FORMAT_MJPEG:
+		return this.format_mjpeg;
+	  case this.RS2_FORMAT_Y8I:
+		return this.format_y8i;
+	  case this.RS2_FORMAT_Y12I:
+		return this.format_y12i;
+	  case this.RS2_FORMAT_INZI:
+		return this.format_inzi;
+	  case this.RS2_FORMAT_INVI:
+		return this.format_invi;
+	  case this.RS2_FORMAT_W10:
+		return this.format_w10;
     }
   },
 };
